# Optimize `apply_K`