A few low-hanging fruits to optimize `solve_ΩplusK`, which is currently dominated by `apply_K`.

### 1. Optimize allocations in `compute_δρ`
A couple of `.` were missing to fuse array operations together.

Before:
```julia
julia> @time δρ = DFTK.compute_δρ(basis_ref, ψ_upscaled, δψ.δψ, occupation);
  0.422897 seconds (5.26 k allocations: 538.114 MiB, 4.43% gc time)
```
After:
```julia
julia> @time δρ = DFTK.compute_δρ(basis_ref, ψ_upscaled, δψ.δψ, occupation);
  0.328420 seconds (4.61 k allocations: 23.110 MiB)
```

### 2. Optimize allocations in `apply_K`
For each k-point, `apply_K` was allocating three arrays per band: one to hold the IFFT, one to hold `δV .* ψnk_real`, and one to hold the FFT.

The change is to now preallocate a single array to hold the result of the IFFT, and of the multiplication by `δV`.

Before:
```julia
julia> @time x = DFTK.apply_K(basis_ref, δψ.δψ, ψ_upscaled, ρ, occupation);
  0.870054 seconds (6.35 k allocations: 689.433 MiB, 5.23% gc time)
```

After:
```julia
julia> @time x = DFTK.apply_K(basis_ref, δψ.δψ, ψ_upscaled, ρ, occupation);
  0.644315 seconds (5.49 k allocations: 65.835 MiB)
```

### 3. Reuse IFFT of `ψ`
~~Every time $\pmb{K}$ is applied, the IFFT of `ψ` was computed **twice**: once in `compute_δρ` and once in `apply_K` itself. It does not change over the entire `solve_ΩplusK` call, so we can compute it once and reuse it.~~

Reverted because it increases memory requirements by an order of magnitude.