@antoine-levitt 2024-10-09
Nice! I assume this is for a very small test case? For bigger ones everything to do with K should be pretty much negligible.