@antoine-levitt 2024-10-09
The point is that omega scales like the number of bands and kpoints, and K doesn't, so for larger unit cells and/or kpoint sampling you'd get rid of K. But still good to have anyway!